att: software basc em c, trabalho (compartilhando arquivo)
